Updated glslang.

+++ 3rdparty/glslang/glslang/MachineIndependent/preprocessor/PpTokens.cpp
@@ -188,7 +188,7 @@ void TPpContext::TokenStream::putToken(int atom, TPpToken* ppToken)
     // save the numeric value
     if (SaveValue(atom)) {
         const char* n = reinterpret_cast<const char*>(&ppToken->i64val);
-        for (int i = 0; i < sizeof(ppToken->i64val); ++i)
+        for (size_t i = 0; i < sizeof(ppToken->i64val); ++i)
             putSubtoken(*n++);
     }
 }
@@ -238,7 +238,7 @@ int TPpContext::TokenStream::getToken(TParseContextBase& parseContext, TPpToken 
     // get the numeric value
     if (SaveValue(atom)) {
         char* n = reinterpret_cast<char*>(&ppToken->i64val);
-        for (int i = 0; i < sizeof(ppToken->i64val); ++i)
+        for (size_t i = 0; i < sizeof(ppToken->i64val); ++i)
             *n++ = getSubtoken();
     }
